Create README.md
Paper is aimed to develop a model using Convolutional Neural Network (CNN) for better classification of
ASL alphabet. Difficulties of our work are in using similar
finger positions for different letters. As a way of reducing
these difficulties augmentation of data and choosing good
hyper-parameters will be used.

#DATA
The data set for this work is taken from KAGGLE, which is called ASL Alphabet. It is available data set, here is
a link: https://www.kaggle.com/grassknoted/asl-alphabet.The data set has 87000 images with 3000 images per category. 
It is separated in 29 folders representing 29 classes. In29 classes, 26 of it are for the letters from A to Z and other 
3 classes for DELETE, SPACE, and NOTHING.These dataset are helpful in real time classifications. There is also, folder 
for test set which consists 29 images. It is given as real world test images, only for prediction.